refactor: split minimap and konoha town modules
