+++ src/hooks/minimapConstants.js
@@ -1,0 +1,16 @@
+export const MINIMAP_DRAW_ROADS = true;
+export const MINIMAP_DRAW_RIVER = true;
+export const MINIMAP_DRAW_DISTRICTS = true;
+export const MINIMAP_ROAD_ALPHA = 0.9;
+export const MINIMAP_W_PRIMARY = 10.0;
+export const MINIMAP_W_SECONDARY = 7.0;
+export const MINIMAP_W_TERTIARY = 4.0;
+
+export const MINIMAP_DRAW_WALLS = true;
+export const MINIMAP_WALL_ALPHA = 0.9;
+export const MINIMAP_WALL_STROKE_SCALE = 2.0;
+export const MINIMAP_WALL_COLOR = '#bfc0c2';
+
+export const MINIMAP_SHOW_DISTRICT = true;
+export const MINIMAP_SHOW_ROAD = true;
+export const MINIMAP_ROAD_MAX_DISTANCE = 18;
